refactor: fix imports due to moving indexes
apache/gravitino#8589
